+++ docs/make.jl
@@ -15,7 +15,8 @@ makedocs(;
         assets = String[]
         ),
     #format = Documenter.HTML(; prettyurls = get(ENV, "CI", nothing) == "true"),
-    pages = ["Domains" => "functions.md",
+    pages = ["Home" => "index.md",
+        "Domains" => "functions.md",
         "Index" => "api.md",
         "News" => "news.md"]
     )